pyproject.toml
[tool.poetry]
name = "everythingev"
version = "0.1.0"
description = "SIH 2020 project"
authors = ["Praful932 <praful.mohanan@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.6"
django = "3.1.14"
django-allauth = "0.41.0"
django-crispy-forms = "^1.14.0"
flake8 = "^6.0.0"
geopy = "^2.3.0"
pillow = "^9.3.0"
pre-commit = "^2.20.0"
django-import-export = "2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
